♻️ Rename: add chatlist folder
