+++ sws/src/chat/components/chatlist/PreChat.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import theme from '../../styles/theme';
+import theme from '../../../styles/theme';
 
 export default function PreChat() {
   return (
